+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 ### Bug fixes
   - Fix loading states causing nested LiveViews to be removed during live navigation
 
+### Enhancements
+  - Add `live_title_tag/2` helper for automatic prefix/postfix on `@page_title` updates
+
 ## 0.9.0 (2020-03-08)
 
 ### Bug fixes
@@ -19,7 +22,6 @@
   - Make flash optional (i.e. LiveView will still work if you don't `fetch_flash` before)
 
 ### Enhancements
-  - Add `live_title_tag/2` helper for automatic prefix/postfix on `@page_title` updates
   - Raise if `:flash` is given as an assign
   - Support user-defined metadata in router
   - Allow the router to be accessed as `socket.router`